Update @types/react, correctly declare react-spring/addons
