+++ package.json
@@ -5,13 +5,13 @@
   "main": "web.cjs.js",
   "module": "web.js",
   "react-native": "native.js",
-  "typings": "index.d.ts",
+  "types": "types/index.d.ts",
   "private": true,
   "sideEffects": false,
   "scripts": {
     "prebuild": "rimraf dist",
     "build": "npm-run-all --parallel copy rollup",
-    "copy": "copyfiles package.json readme.md LICENSE.md index.d.ts dist && json -I -f dist/package.json -e \"this.private=false; this.devDependencies=undefined; this.scripts=undefined; this.husky=undefined; this.prettier=undefined; this.jest=undefined; this['lint-staged'] = undefined;\"",
+    "copy": "copyfiles package.json readme.md LICENSE.md types dist && json -I -f dist/package.json -e \"this.private=false; this.devDependencies=undefined; this.scripts=undefined; this.husky=undefined; this.prettier=undefined; this.jest=undefined; this['lint-staged'] = undefined;\"",
     "rollup": "rollup -c",
     "prepare": "npm run build",
     "test": "jest",
@@ -72,7 +72,7 @@
     "@babel/preset-react": "^7.0.0",
     "@storybook/addon-knobs": "^3.4.11",
     "@storybook/react": "^3.4.11",
-    "@types/react": "^16.7.8",
+    "@types/react": "^16.7.17",
     "babel-core": "7.0.0-bridge.0",
     "babel-jest": "^23.6.0",
     "babel-plugin-react-docgen": "^2.0.0",
