+++ tsconfig.json
@@ -8,7 +8,8 @@
     "noFallthroughCasesInSwitch": true,
     "noEmitOnError": true,
     "forceConsistentCasingInFileNames": true,
-    "lib": ["es2016", "dom"]
+    "lib": ["es2016", "dom"],
+    "skipLibCheck": false
   },
-  "include": ["tests/**/*.tsx"]
+  "include": ["tests/**/*.tsx", "types/**/*.ts"]
 }
