+++ types/index.d.ts
@@ -1,0 +1,6 @@
+/// <reference types="./react-spring"/>
+/// <reference types="./addons"/>
+
+declare module 'react-spring/web' {
+  export * from 'react-spring'
+}
